v0.5.0: 7-tier TS-logic roadmap — bugs, security, architecture, perf, config, tests, features
Tier 1 - Bugs (8 fixes):
- mode_manager deadline uses monotonic elapsed-time tracking
- router checks request_user_mode return, returns error on timeout
- calc tool supports modulo and floor division operators
- VaultAdapter uses config vault root, not raw source as path
- Inference fallback returns consistent tuple shape
- Shutdown consolidation configurable, not hour-gated
- ThrottlePolicy uses float consistently
- RLock reentrancy in graph load() documented

Tier 2 - Security (5 hardening items):
- core/path_sandbox.py reused by markdown, vault, file_read
- RSS XML bomb protection with size cap + optional defusedxml
- FileReadTool max_bytes limit and pinned base_dir
- Ollama base_url configurable via config.yaml
- Dashboard defaults to 127.0.0.1, warns without auth token

Tier 3 - Architecture (4 splits):
- WaveCycleRunner extracted to core/graph/wave_runner.py
- BoggersRuntime split: AutonomousLoopMixin + SelfImprovementMixin
- Dashboard lazy runtime via get_runtime() with double-checked lock
- Runtime reduced from 1100 to 550 lines

Tier 4 - Performance (4 optimizations):
- Consolidation topic-bucketed O(k) vs O(n^2)
- get_activated_subgraph uses heapq.nlargest
- Adapter cache thread-safe with Lock
- Guardrails reads under graph lock

Tier 5 - Config/DX (5 items):
- config_schema strict mode + env BOGGERS_CONFIG_STRICT
- RuntimeConfig wired for graph_backend, sqlite_path, snapshot_dir
- pytest ini_options, .pre-commit-config.yaml, Makefile

Tier 6 - Tests/CI (7 new test files + strengthened):
- test_config_loader, test_plugins, test_sqlite_backend, test_rules_engine
- test_inference_router, test_path_sandbox, test_wave_runner
- Strengthened consolidation, events, wave assertions
- CI: pip cache, mypy blocking, cov-fail-under 60

Tier 7 - Features (8 new modules):
- adapters/http_client.py with retry + exponential backoff
- core/graph/operations.py: BFS subgraph, batch_add, components, range filter
- tools: web_search, datetime, unit_convert
- Dashboard /health/live and /health/ready endpoints
- 216 tests total, all passing

Made-with: Cursor

+++ entities/inference_router.py
@@ -13,7 +13,7 @@ class SynthesisProtocol(Protocol):
 
 @dataclass(slots=True)
 class ThrottlePolicy:
-    min_interval_seconds: int = 60
+    min_interval_seconds: float = 60.0
 
 
 class InferenceRouter:
